% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_imputation.R
\name{compare_imputation}
\alias{compare_imputation}
\title{Compare Imputation Methods}
\usage{
compare_imputation(data)
}
\arguments{
\item{data}{A data frame with missing values.}
}
\value{
A summary of the comparison.
}
\description{
This function compares different imputation methods in terms of computational time.
}
\examples{
data <- data.frame(var1 = c(1, NA, 2, 4, 5),var2 = c(2, 3, 4, NA, 7),var3 = c(2, 3, 4, 5, NA))
compare_imputation(data)
}
