% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{knn_impute}
\alias{knn_impute}
\title{Impute Missing Values Using k-Nearest Neighbors (kNN)}
\usage{
knn_impute(data, k = 3)
}
\arguments{
\item{data}{A matrix or data frame containing missing values (NAs) to be imputed.}

\item{k}{The number of nearest neighbors to consider for imputation. Default is 5.}
}
\value{
A matrix with missing values imputed using the kNN algorithm.
}
\description{
This function uses the kNN() function from the VIM package to impute missing values in a given matrix.
}
\examples{
# Create a sample matrix with missing values
mat <- matrix(c(1, 2, NA, 4, 5, NA, 7, 8, 9), nrow = 3, byrow = TRUE)
imputed_matrix <- knn_impute(mat, k = 3)
print(imputed_matrix)

}
