% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest_impute.R
\name{random_forest_impute}
\alias{random_forest_impute}
\title{Random Forest-based Missing Value Imputation}
\usage{
random_forest_impute(data, ntree = 100, max_iter = 10)
}
\arguments{
\item{data}{A data frame with missing values (NA).}

\item{ntree}{Number of trees in the forest. Default is 100.}

\item{max_iter}{Maximum number of iterations for imputation. Default is 10.}
}
\value{
A complete data frame with missing values imputed.
}
\description{
This function imputes missing values using Random Forest, inspired by the missForest algorithm.
}
\examples{
data <- data.frame(var1 = c(1, 2, NA, 4, 5),var2 = c(NA, 3, 4, NA, 6),var3 = c(1, NA, NA, 4, 5))
random_forest_impute(data, ntree = 50)
}
