\name{SA24204177-package}
\alias{SA24204177-package}
\docType{package}
\title{missdata solver }
\description{This package provides robust missing data imputation methods including EM algorithm (Rcpp-optimized) and Random Forest imputation.Users can compare their performance in terms of speed and accuracy. }
\author{
Yifei Zhu
Maintainer: Yifei Zhu <zyfshanshan@mail.ustc.edu.cn>
}
\keyword{ package }
